# api/talos.openapi-v2.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ory Talos API",
    "description": "Ory Talos is a high-performance API key management service. It handles the full API key lifecycle: issuing keys, verifying them at low latency, deriving short-lived tokens (JWT and Macaroon), and revoking access.\n\n`Admin*`-prefixed RPCs require admin authentication. The bare `RevokeAPIKey` RPC is authenticated by proof of possession (the raw credential secret).",
    "version": "v2alpha1"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v2alpha1/admin/importedApiKeys": {
      "get": {
        "summary": "List Imported API Keys",
        "description": "Lists all imported keys with filtering. Returns imported keys only (not\nissued keys). Supports pagination and AIP-160 filter expressions.\n\n```http\nGET /v2alpha1/admin/importedApiKeys?page_size=50\u0026filter=status%3DKEY_STATUS_ACTIVE\n```",
        "operationId": "AdminListImportedAPIKeys",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ListImportedAPIKeysResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "page_size",
            "description": "Number of items per page (default: 50, max: 1000)",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_token",
            "description": "Cursor token for pagination (OPTIONAL)",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "filter",
            "description": "filter is an AIP-160 expression. Indexed fields (efficient at any scale):\n  actor_id, status. Other fields are not indexed and may be rejected.\nExamples:\n  actor_id=\"user_123\"\n  status=KEY_STATUS_ACTIVE\n  actor_id=\"user_123\" AND status=KEY_STATUS_ACTIVE",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "APIKeys"
        ]
      },
      "post": {
        "summary": "Import API Key",
        "description": "Imports an external API key into the system. Allows importing keys from\nlegacy systems or external providers. The raw key is hashed and stored\nsecurely (HMAC). Imported keys support token derivation (JWT/Macaroon)\nlike issued keys.\n\n```http\nPOST /v2alpha1/admin/importedApiKeys\n{\n  \"raw_key\": \"sk_live_abc123xyz\",\n  \"name\": \"Imported Stripe Key\",\n  \"actor_id\": \"user_123\"\n}\n```",
        "operationId": "AdminImportAPIKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ImportedAPIKey"
            }
          },
          "201": {
            "description": "API key imported successfully.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Example:\n  {\n    \"raw_key\": \"sk_live_abc123xyz789\",\n    \"name\": \"Stripe Production Key\",\n    \"actor_id\": \"payment-processor\",\n    \"scopes\": [\"read\", \"write\"],\n    \"ttl\": \"8760h\",  // 1 year (also accepts: 31536000s)\n    \"metadata\": {\"source\": \"stripe\", \"environment\": \"production\"}\n  }",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ImportAPIKeyRequest"
            }
          }
        ],
        "tags": [
          "APIKeys"
        ]
      }
    },
...
    "/v2alpha1/admin/issuedApiKeys": {
      "get": {
        "summary": "List Issued API Keys",
        "description": "Lists issued API keys with optional filtering. Supports cursor-based\npagination and AIP-160 filter expressions. Returns only issued\n(generated) API keys; use ListImportedAPIKeys for imported keys.\n\n```http\nGET /v2alpha1/admin/issuedApiKeys?page_size=50\u0026filter=actor_id%3D%22user_123%22\n```",
        "operationId": "AdminListIssuedAPIKeys",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ListIssuedAPIKeysResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "page_size",
            "description": "Number of items per page (default: 50, max: 1000)",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_token",
            "description": "Cursor token for pagination",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "filter",
            "description": "filter is an AIP-160 expression. Indexed fields (efficient at any scale):\n  actor_id, status. Other fields are not indexed and may be rejected.\nExamples:\n  actor_id=\"user_123\"\n  status=KEY_STATUS_ACTIVE\n  actor_id=\"user_123\" AND status=KEY_STATUS_ACTIVE",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "APIKeys"
        ]
      },
      "post": {
        "summary": "Issue API Key",
        "description": "Creates a new API key for a given actor. The secret is returned only once\nin the response and cannot be retrieved later. Keys can be scoped with\nspecific permissions and have optional expiration.\n\n```http\nPOST /v2alpha1/admin/issuedApiKeys\n{\n  \"name\": \"production-service\",\n  \"actor_id\": \"user_123\",\n  \"scopes\": [\"read\", \"write\"],\n  \"ttl\": \"8760h\"\n}\n```",
        "operationId": "AdminIssueAPIKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/IssueAPIKeyResponse"
            }
          },
          "201": {
            "description": "API key issued successfully.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IssueAPIKeyRequest"
            }
          }
        ],
        "tags": [
          "APIKeys"
        ]
      }
    },
...
    "/v2alpha1/apiKeys:selfRevoke": {
      "post": {
        "summary": "Revoke API Key (self-service)",
        "description": "Proof-of-possession variant of revocation. Lives alongside `AdminRevokeAPIKey`\nin this service; the `Self*` prefix on the request/response messages\ndisambiguates from the admin variant's `RevokeAPIKeyRequest`.\n\nAllows an API key holder to revoke their own key. The caller must provide\nthe full API key secret as proof of possession. Supports issued API keys\nand imported keys. JWT and macaroon tokens cannot be self-revoked (they\nare stateless).\n\nThe PRIVILEGE_WITHDRAWN reason is not allowed for self-revocation\n(admin-only).\n\n```http\nPOST /v2alpha1/apiKeys:selfRevoke\n{\n  \"credential\": \"sk_live_abc123...\",\n  \"reason\": \"REVOCATION_REASON_KEY_COMPROMISE\"\n}\n```",
        "operationId": "RevokeAPIKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/SelfRevokeAPIKeyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "SelfRevokeAPIKeyRequest allows an API key holder to revoke their own key\nby providing the full key secret as proof of possession.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SelfRevokeAPIKeyRequest"
            }
          }
        ],
        "tags": [
          "APIKeys"
        ]
      }
    },
...
  },
  "definitions": {
...
    "Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "IPRestriction": {
      "type": "object",
      "properties": {
        "allowed_cidrs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "allowed_cidrs is a list of CIDR ranges that are allowed to use this key.\nSupports both IPv4 (e.g., \"10.0.0.0/8\") and IPv6 (e.g., \"2001:db8::/32\").\nIf empty, all IPs are allowed (no restriction)."
        }
      },
      "description": "IPRestriction defines IP-based access controls for an API key.\nWhen allowed_cidrs is non-empty, only requests from IPs matching at least one\nCIDR range are permitted. Empty allowed_cidrs means no IP restriction (all IPs allowed).\nIP restrictions apply to root API key and imported key verification only;\nderived tokens (JWT/macaroon) are stateless and not subject to IP checks."
    },
    "ImportAPIKeyRequest": {
      "type": "object",
      "properties": {
        "raw_key": {
          "type": "string",
          "title": "The actual key string to import (REQUIRED)"
        },
        "name": {
          "type": "string",
          "title": "Human-readable name (REQUIRED)"
        },
        "actor_id": {
          "type": "string",
          "description": "actor_id is the identifier of the entity that owns this imported key.\nRequired so every imported key is traceable to an actor for revocation and\naudit queries."
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Scopes for the imported key (OPTIONAL)"
        },
        "ttl": {
          "type": "string",
          "description": "ttl sets the expiry as a duration from now. Encoded as a\ngoogle.protobuf.Duration (string ending in \"s\", e.g. \"3600s\").\nAccepted bounds: 1s to 315360000s (~10 years). If unset or zero, the\nproject default TTL applies.\nFor convenience, the server also accepts Go-style duration strings\n(\"24h\", \"30m\", \"1h30m\") and an extended unit set (\"1d\", \"1w\", \"1mo\",\n\"1y\"; approximations: 1mo = 30d, 1y = 365d). Clients should prefer\nthe standard Duration encoding for portability."
        },
        "metadata": {
          "type": "object",
          "description": "metadata is a free-form JSON object for caller-defined attributes (e.g.,\nsource, environment, tags). Values may be strings, numbers, booleans,\narrays, objects, or null. Total serialized size is capped at 4KB.\nAIP-148 metadata field."
        },
        "rate_limit_policy": {
          "$ref": "#/definitions/RateLimitPolicy"
        },
        "ip_restriction": {
          "$ref": "#/definitions/IPRestriction"
        },
        "request_id": {
          "type": "string",
          "title": "Client-controlled idempotency key (AIP-155)"
        },
        "visibility": {
          "$ref": "#/definitions/KeyVisibility"
        }
      },
      "description": "Example:\n  {\n    \"raw_key\": \"sk_live_abc123xyz789\",\n    \"name\": \"Stripe Production Key\",\n    \"actor_id\": \"payment-processor\",\n    \"scopes\": [\"read\", \"write\"],\n    \"ttl\": \"8760h\",  // 1 year (also accepts: 31536000s)\n    \"metadata\": {\"source\": \"stripe\", \"environment\": \"production\"}\n  }",
      "title": "ImportAPIKeyRequest imports an external HMAC-based API key"
    },
    "ImportedAPIKey": {
      "type": "object",
      "properties": {
        "key_id": {
          "type": "string",
          "title": "SHA-512/256 hash of credential"
        },
        "name": {
          "type": "string"
        },
        "actor_id": {
          "type": "string"
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "metadata": {
          "type": "object",
          "description": "metadata is a free-form JSON object for caller-defined attributes\n(e.g., source, environment, tags). Values may be strings, numbers,\nbooleans, arrays, objects, or null. Total serialized size is capped\nat 4KB. AIP-148 metadata field."
        },
        "status": {
          "$ref": "#/definitions/KeyStatus"
        },
        "create_time": {
          "type": "string",
          "format": "date-time"
        },
        "update_time": {
          "type": "string",
          "format": "date-time"
        },
        "expire_time": {
          "type": "string",
          "format": "date-time"
        },
        "last_used_time": {
          "type": "string",
          "format": "date-time"
        },
        "rate_limit_policy": {
          "$ref": "#/definitions/RateLimitPolicy"
        },
        "revocation_reason": {
          "$ref": "#/definitions/RevocationReason"
        },
        "revocation_description": {
          "type": "string",
          "description": "revocation_description provides free-form context for a revocation.\nOnly set when revocation_reason is PRIVILEGE_WITHDRAWN.\nJSON API change: field was formerly revocation_reason_text. Field number 13\nis unchanged so the change is wire-compatible for binary proto encoding."
        },
        "ip_restriction": {
          "$ref": "#/definitions/IPRestriction"
        },
        "visibility": {
          "$ref": "#/definitions/KeyVisibility"
        }
      },
      "description": "ImportedAPIKey represents an API key imported from an external system.\nThe raw key is hashed (SHA-512/256) and stored. The original key is never retained."
    },
    "IssueAPIKeyRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "actor_id": {
          "type": "string"
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "ttl": {
          "type": "string",
          "description": "ttl sets the expiry as a duration from now. Encoded as a\ngoogle.protobuf.Duration (string ending in \"s\", e.g. \"3600s\").\nAccepted bounds: 1s to 315360000s (~10 years). If unset or zero, the\nproject default TTL applies.\nFor convenience, the server also accepts Go-style duration strings\n(\"24h\", \"30m\", \"1h30m\") and an extended unit set (\"1d\", \"1w\", \"1mo\",\n\"1y\"; approximations: 1mo = 30d, 1y = 365d). Clients should prefer\nthe standard Duration encoding for portability."
        },
        "metadata": {
          "type": "object",
          "description": "metadata is a free-form JSON object for caller-defined attributes\n(e.g., source, environment, tags). Values may be strings, numbers,\nbooleans, arrays, objects, or null. Total serialized size is capped\nat 4KB. AIP-148 metadata field."
        },
        "rate_limit_policy": {
          "$ref": "#/definitions/RateLimitPolicy"
        },
        "request_id": {
          "type": "string",
          "title": "Client-controlled idempotency key (AIP-155)"
        },
        "ip_restriction": {
          "$ref": "#/definitions/IPRestriction"
        },
        "visibility": {
          "$ref": "#/definitions/KeyVisibility"
        }
      }
    },
    "IssueAPIKeyResponse": {
      "type": "object",
      "properties": {
        "issued_api_key": {
          "$ref": "#/definitions/IssuedAPIKey"
        },
        "secret": {
          "type": "string",
          "title": "Only returned on creation"
        }
      }
    },
...
    "KeyStatus": {
      "type": "string",
      "enum": [
        "KEY_STATUS_UNSPECIFIED",
        "KEY_STATUS_ACTIVE",
        "KEY_STATUS_REVOKED",
        "KEY_STATUS_EXPIRED"
      ],
      "default": "KEY_STATUS_UNSPECIFIED",
      "description": "KeyStatus represents the lifecycle state of an API key.\n\n - KEY_STATUS_UNSPECIFIED: Default zero value. Never returned by the server. Treated as ACTIVE for\nbackward compatibility but should not be relied on.\n - KEY_STATUS_ACTIVE: The key is valid and can be used to authenticate.\n - KEY_STATUS_REVOKED: The key was revoked. Verification fails with VERIFICATION_ERROR_REVOKED.\nSee revocation_reason for the cause.\n - KEY_STATUS_EXPIRED: The key passed its expire_time. Verification fails with\nVERIFICATION_ERROR_EXPIRED. The transition is computed at read time and\nnot persisted."
    },
    "KeyVisibility": {
      "type": "string",
      "enum": [
        "KEY_VISIBILITY_UNSPECIFIED",
        "KEY_VISIBILITY_SECRET",
        "KEY_VISIBILITY_PUBLIC"
      ],
      "default": "KEY_VISIBILITY_UNSPECIFIED",
      "description": "KeyVisibility distinguishes public (client-safe) keys from secret (server-only) keys.\nPublic keys use a different configurable prefix for visual distinction.\nBoth types share the same scope/permission system — visibility is about exposure safety.\n\n - KEY_VISIBILITY_UNSPECIFIED: Treated as SECRET"
    },
    "ListImportedAPIKeysResponse": {
      "type": "object",
      "properties": {
        "imported_api_keys": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ImportedAPIKey"
          },
          "title": "List of imported keys"
        },
        "next_page_token": {
          "type": "string",
          "title": "Token for fetching the next page (empty if no more results)"
        }
      },
      "title": "ListImportedAPIKeysResponse returns paginated list of imported keys"
    },
    "ListIssuedAPIKeysResponse": {
      "type": "object",
      "properties": {
        "issued_api_keys": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/IssuedAPIKey"
          }
        },
        "next_page_token": {
          "type": "string"
        }
      }
    },
...
    "RateLimitPolicy": {
      "type": "object",
      "properties": {
        "quota": {
          "type": "string",
          "format": "int64",
          "description": "quota is the number of requests allowed per window."
        },
        "window": {
          "type": "string",
          "description": "window is the time window for the quota.\nCommon values: 60s (1 minute), 3600s (1 hour), 86400s (1 day)."
        },
        "unit": {
          "type": "string",
          "title": "unit describes the quota unit type.\nDefault: \"requests\"\nOther valid values per IETF spec: \"content-bytes\", \"concurrent-requests\""
        }
      },
      "description": "RateLimitPolicy describes the rate limit policy for an API key.\n\nIn OSS mode, this policy is informational and meant to be consumed by\nupstream gateways (Envoy, Cloudflare, etc.) for enforcement.\nIn commercial mode, Talos enforces rate limits using in-memory or Redis backends,\nboth powered by the GCRA (Generic Cell Rate Algorithm).\n\nCompliant with draft-ietf-httpapi-ratelimit-headers-10."
    },
    "RevocationReason": {
      "type": "string",
      "enum": [
        "REVOCATION_REASON_UNSPECIFIED",
        "REVOCATION_REASON_KEY_COMPROMISE",
        "REVOCATION_REASON_AFFILIATION_CHANGED",
        "REVOCATION_REASON_SUPERSEDED",
        "REVOCATION_REASON_PRIVILEGE_WITHDRAWN"
      ],
      "default": "REVOCATION_REASON_UNSPECIFIED",
      "description": "RevocationReason provides structured revocation reasons inspired by RFC 5280.\nUsed in both admin and self-revocation flows.\n\n - REVOCATION_REASON_UNSPECIFIED: Default zero value. Use a specific reason; UNSPECIFIED is rejected by\nadmin and self-revocation endpoints.\n - REVOCATION_REASON_KEY_COMPROMISE: The key was leaked or believed to be in the hands of an unauthorized\nparty.\n - REVOCATION_REASON_AFFILIATION_CHANGED: The owning actor's relationship with the issuer changed (e.g., role\nchange, departure).\n - REVOCATION_REASON_SUPERSEDED: A new key has replaced this one as part of a rotation.\n - REVOCATION_REASON_PRIVILEGE_WITHDRAWN: Admin-only. The actor's privilege to use this key was withdrawn by\nan operator. Self-revocation requests using this reason are rejected\nwith InvalidArgument. Pair with `description` on RevokeAPIKeyRequest\nto record the operator-supplied justification."
    },
...
    "SelfRevokeAPIKeyRequest": {
      "type": "object",
      "properties": {
        "credential": {
          "type": "string",
          "title": "Full API key secret or imported key (REQUIRED)"
        },
        "reason": {
          "$ref": "#/definitions/RevocationReason",
          "description": "reason explains why the holder is revoking the key. Must be one of\nKEY_COMPROMISE, AFFILIATION_CHANGED, or SUPERSEDED. PRIVILEGE_WITHDRAWN\nis admin-only and is rejected here with InvalidArgument."
        }
      },
      "description": "SelfRevokeAPIKeyRequest allows an API key holder to revoke their own key\nby providing the full key secret as proof of possession."
    },
    "SelfRevokeAPIKeyResponse": {
      "type": "object",
      "description": "SelfRevokeAPIKeyResponse is returned on successful self-revocation.\n\nEmpty on success"
    },
    "Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Any"
          }
        }
      }
    },
...
  }
}